+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/ClassAndEnum.cs
@@ -1,0 +1,10 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public class ClassNextToEnum
+    {
+    }
+
+    public enum Enum
+    {
+    }
+}