+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/SingleClass.cs
@@ -1,0 +1,10 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public class SingleClass
+    {
+        // Prove that nested classes are not considered as top level types:
+        public class Nested
+        {
+        }
+    }
+}