+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/SingleClassWithWrongName.cs
@@ -1,0 +1,6 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public class SingleClassWithWrongNameX
+    {
+    }
+}