+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/TwoClasses.cs
@@ -1,0 +1,10 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public class TwoClassesA
+    {
+    }
+
+    public class TwoClassesB
+    {
+    }
+}