+++ History.md
@@ -1,5 +1,6 @@
 ### vNext ###
-- Add "To-do items count" (see options page)
+- Added "To-do items count" (see options page)
+- Added "Mismatch between type and file name" warning
 
 ### 0.7.0 ###
 - Added provider for "Copy Fully-qualified name" which returns <TypeShortName>.<MemberShortName>